0.103.9: lint cleanup — remove empty catch bodies and unnecessary type assertions

+++ src/stash-crypto.ts
@@ -85,7 +85,7 @@ async function deriveKeyBytes(password: string, salt: Uint8Array, kdf: KdfSpec):
       hashLength: KEY_LEN,
       outputType: "binary",
     });
-    return out as Uint8Array;
+    return out;
   }
   if (kdf.id === KDF_PBKDF2) {
     const baseKey = await crypto.subtle.importKey(
